requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
langchain-community
langchain-pinecone
numpy
pandas
python-dotenv
langchain_openai 
langchain_core
streamlit
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox